Player Inertia

+++ Unity/Assets/Scripts/Data/Player.cs
@@ -10,7 +10,7 @@ public struct Player
     public Vector2 position;
     public long lastShootStep;
     public bool isGameOver;
-    public long velocity;
+    public float velocity;
     //public int playerID;
 
 }
